Organize Transactions into its own package
